# src/core/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "uxwVend API",
    "description": "Auto-generated API surface for uxwVend. Core endpoints are hand-curated; module endpoints are derived from each installed module's `module.json` manifest.",
    "version": "1.0.0",
    "contact": {
      "name": "uxwVend",
      "url": "https://github.com/siracozmen01/uxwVend"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register new user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "description": "Create a new user account with email, username and password; validates input with Zod and hashes credentials with bcrypt.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "username",
                  "password",
                  "confirmPassword"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/profile": {
      "get": {
        "summary": "Get current user profile",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Profile data"
          }
        },
        "description": "Return the authenticated user's profile, including display fields and preferences.",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update current user profile",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "description": "Update the authenticated user's own profile fields (name, bio, avatar, preferences).",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}